Took 6.73 seconds (891.49% speed)

+++ src/nes/ppu/ppu.ts
@@ -1,4 +1,4 @@
-import { assertInRange, assertUint8, uint16, uint8, uint8Reverse } from '../num'
+import { assertInRange, assertUint8, uint16, uint8 } from '../num'
 import { NMI } from '../cpu/nmi'
 import * as Color from './color'
 import { Logger } from '../logger'
@@ -461,7 +461,7 @@ export class PPU {
     // Stores color index 0-63, or -1 for no pixel.
     // If pixel exists the bit 64 (1<<6) represents its priority.
     // If the pixel is from sprite 0, the bit 128 (1<<7) is set for sprite zero hit detection.
-    private spriteLineBuffer = new Int32Array(WIDTH)
+    private spriteLineBuffer = new Int16Array(WIDTH)
     // Computes sprite pixels for scanline and stores the result to spriteLineBuffer.
     //
     // References
